+++ babel.config.js
@@ -1,0 +1,11 @@
+module.exports = {
+  presets: [
+    [
+      '@babel/preset-env',
+      {
+        targets: { electron: 5 },
+      },
+    ],
+    '@babel/preset-typescript',
+  ],
+}