+++ src/types.js
@@ -1,6 +1,0 @@
-/* @flow */
-
-export type Dependency = {
-  url: ?string,
-  name: string,
-}