+++ src/types.ts
@@ -1,0 +1,8 @@
+export interface Dependency {
+  name: string
+  minimumVersion?: string | null
+}
+
+export interface DependencyResolved extends Dependency {
+  directory: string | null
+}